+++ src/app/core/doctor-schedule/schedule/schedule.component.ts
@@ -52,8 +52,48 @@ export class ScheduleComponent implements OnInit{
   this.libreta=dato;
     });}
     eliminarPersona(id: number) {
-      this.libretaService.eliminarPersona(id).subscribe(() => {
-        this.obtenerPersona(); // Para actualizar la lista después de la eliminación
+      // Mostrar la alerta personalizada
+      this.mostrarAlerta('¿Desea eliminar esta persona?').then((confirmacion) => {
+        // Si el usuario hace clic en "Aceptar" en la alerta personalizada
+        if (confirmacion) {
+          this.libretaService.eliminarPersona(id).subscribe(() => {
+            // Actualizar la lista después de eliminar
+            this.obtenerPersona();
+    
+            // Recargar la página después de la eliminación
+           location.reload();
+          });
+        }
+      });
+    }
+    
+    recargarPagina() {
+      // Recargar la página
+      location.reload();
+    } 
+    mostrarAlerta(mensaje: string): Promise<boolean> {
+      return new Promise<boolean>((resolve) => {
+        const customAlert = document.getElementById('customAlert') as HTMLElement;
+        const alertMessage = document.getElementById('alertMessage') as HTMLElement;
+        const confirmButton = document.getElementById('confirmButton') as HTMLButtonElement;
+        const cancelButton = document.getElementById('cancelButton') as HTMLButtonElement;
+    
+        // Configurar el mensaje
+        alertMessage.innerText = mensaje;
+    
+        // Mostrar la alerta
+        customAlert.style.display = 'flex';
+    
+        // Configurar eventos de los botones
+        confirmButton.onclick = () => {
+          customAlert.style.display = 'none';
+          resolve(true);
+        };
+    
+        cancelButton.onclick = () => {
+          customAlert.style.display = 'none';
+          resolve(false);
+        };
       });
     }
