+++ src/app/core/staff/staff-list/staff-list.component.ts
@@ -49,7 +49,10 @@ export class StaffListComponent implements OnInit{
     //aqui solo dirige ala pagina de actualizar maquina
     this.router.navigate([routes.addLeave,{id}]);
   }
-  
+  recargarPagina() {
+    // Recargar la página
+    location.reload();
+  } 
   obtenerPersona(){
     this.antropometriaService.obtenerListaPersona().subscribe(dato => {
   this.antropometria=dato;
@@ -64,11 +67,46 @@ export class StaffListComponent implements OnInit{
   // console.log(dato)
   //   });}
 
-    eliminarPersona(id: number) {
-      this.antropometriaService.eliminarPersona(id).subscribe(() => {
-        this.obtenerPersona(); // Para actualizar la lista después de la eliminación
-      });
-    }
+  eliminarPersona(id: number) {
+    // Mostrar la alerta personalizada
+    this.mostrarAlerta('¿Desea eliminar esta persona?').then((confirmacion) => {
+      // Si el usuario hace clic en "Aceptar" en la alerta personalizada
+      if (confirmacion) {
+        this.antropometriaService.eliminarPersona(id).subscribe(() => {
+          // Actualizar la lista después de eliminar
+          this.obtenerPersona();
+  
+          // Recargar la página después de la eliminación
+         location.reload();
+        });
+      }
+    });
+  }
+  mostrarAlerta(mensaje: string): Promise<boolean> {
+    return new Promise<boolean>((resolve) => {
+      const customAlert = document.getElementById('customAlert') as HTMLElement;
+      const alertMessage = document.getElementById('alertMessage') as HTMLElement;
+      const confirmButton = document.getElementById('confirmButton') as HTMLButtonElement;
+      const cancelButton = document.getElementById('cancelButton') as HTMLButtonElement;
+  
+      // Configurar el mensaje
+      alertMessage.innerText = mensaje;
+  
+      // Mostrar la alerta
+      customAlert.style.display = 'flex';
+  
+      // Configurar eventos de los botones
+      confirmButton.onclick = () => {
+        customAlert.style.display = 'none';
+        resolve(true);
+      };
+  
+      cancelButton.onclick = () => {
+        customAlert.style.display = 'none';
+        resolve(false);
+      };
+    });
+  }
   
   
   private getTableData(): void {
